% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseshootS-package.R
\docType{package}
\name{sparseshootS-package}
\alias{sparseshootS}
\alias{sparseshootS-package}
\title{sparseshootS: Sparse Shooting S-Estimator for Robust High-Dimensional Regression}
\description{
Computes the sparse shooting S-estimator and the shooting S-estimator for robust regression in high-dimensional settings with cellwise outliers. Implements the method described in Bottmer, Croux, and Wilms (2022), "Sparse regression for large data sets with outliers", Econometrics and Statistics.
}
\author{
\strong{Maintainer}: Ines Wilms \email{i.wilms@maastrichtuniversity.nl}

Authors:
\itemize{
  \item Ines Wilms \email{i.wilms@maastrichtuniversity.nl}
  \item Lea Bottmer
  \item Christophe Croux
}

}
\keyword{internal}
